pyproject.toml
[tool.poetry]
name = "flight-backend"
version = "0.1.0"
description = ""
authors = ["MaHDiaLaGaB <elmahdi.alagab@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
sqlalchemy = "^2.0.1"
alembic = "^1.9.2"
black = "^23.1.0"
pytest = "^7.2.1"
requests = "^2.28.2"
httpx = "^0.23.3"
python-dotenv = "^0.21.1"
mypy = "^1.0.0"
sentry-sdk = "^1.15.0"
amadeus = "^8.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
